back-end/requirements.txt
Flask
Flask-Cors
Pillow
numpy
scikit-image